max/min need to be Double now also for numberParam
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@20494 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/jspwrap/NumberParamBean.java
@@ -28,11 +28,11 @@ public class NumberParamBean extends ParamBean<NumberParam> {
         return _param.isInteger();
     }
     
-    public Long getMin() {
+    public Double getMin() {
       return _param.getMin();
     }
     
-    public Long getMax() {
+    public Double getMax() {
       return _param.getMax();
     }
